Add files via upload
Added explaination regarding results when choosing min_df values and optimizing the hyperparameter alpha.  I originally chose min_df for values between 0.1 and 0.4.  When doing so, I got a best alpha of 50 and a best min_df of 0.1.  The training and test accuracies were 60.6% and 61.5% respectively.  The classifier generalizes to new data very well but the accuracy is lower than the default Naive Bayes classifier.